move more files to models in core, more PR cleanup in IngestController, add scaladoc to RequestSummary
