+++ core/src/main/scala/org/broadinstitute/monster/ingester/core/IngestController.scala
@@ -12,12 +12,7 @@ import doobie.util.transactor.Transactor
 import io.chrisdavenport.log4cats.slf4j.Slf4jLogger
 import org.broadinstitute.monster.ingester.core.ApiError.NotFound
 import org.broadinstitute.monster.ingester.core.IngestController.JobSubmission
-import org.broadinstitute.monster.ingester.core.models.{
-  IngestData,
-  JobData,
-  JobSummary,
-  RequestSummary
-}
+import org.broadinstitute.monster.ingester.core.models.{ApiStatus, IngestData, JobData, JobSummary, RequestSummary}
 import org.broadinstitute.monster.ingester.jade.JadeApiClient
 import org.broadinstitute.monster.ingester.jade.models.{IngestRequest, JobStatus}
 
@@ -216,13 +211,13 @@ class IngestController(dbClient: Transactor[IO], jadeClient: JadeApiClient)(
   def enumerateJobs(requestId: UUID): IO[List[JobSummary]] = {
     checkAndExec(requestId) { rId =>
       List(
-        Fragment.const(
-          s"SELECT jade_id, status, path, table_name, submitted, completed FROM $JobsTable"
-        ),
-        fr"WHERE request_id = $rId"
-      ).combineAll
-        .query[JobSummary]
-        .to[List]
+          Fragment.const(
+            s"SELECT jade_id, status, path, table_name, submitted, completed FROM $JobsTable"
+          ),
+          fr"WHERE request_id = $rId"
+        ).combineAll
+          .query[JobSummary]
+          .to[List]
     }
   }
 
@@ -252,9 +247,7 @@ class IngestController(dbClient: Transactor[IO], jadeClient: JadeApiClient)(
         .to[List]
 
       // hit jade job result endpoint for each job
-      statuses <- Async[ConnectionIO].liftIO(ids.traverse { id =>
-        jadeClient.jobStatus(id)
-      })
+      statuses <- Async[ConnectionIO].liftIO(ids.traverse(jadeClient.jobStatus))
 
       // if status has changed, update, else remain (upsert)
       numUpdated <- statuses.traverse { row =>
