+++ core/src/main/scala/org/broadinstitute/monster/ingester/core/models/ApiError.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.ingester.core
+package org.broadinstitute.monster.ingester.core.models
 
 import java.util.UUID
