+++ core/src/main/scala/org/broadinstitute/monster/ingester/core/models/ApiStatus.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.ingester.core
+package org.broadinstitute.monster.ingester.core.models
 
 // TODO eventually will be a class that corresponds to the information returned when the api status is requested
 case class ApiStatus()
